Refactor code to include donViTinh field in APIDonHangTheoEmail query

+++ server/schemas/index.js
@@ -126,7 +126,7 @@ type ThongKe {
     labels: [String],
     datas: [Float],
     percent: Float,
-    quantity: Int,
+    quantity: Float,
 },
 type Query {
     danhSachSanPham: [SanPham],
